temp/requirements.txt
Django
djangorestframework
uWSGI
